+++ clang/include/clang/Lex/HeaderSearchOptions.h
@@ -38,6 +38,16 @@ enum IncludeDirGroup {
   /// Like Angled, but marks header maps used when building frameworks.
   IndexHeaderMap,
 
+  /// Like Angled, but marks system directories while retaining relative order
+  /// with user directories. This group is intended to match the semantics of
+  /// the MSVC /external:I option.
+  External,
+
+  /// Like External, but searched after other external directories but before
+  /// system directories. This group is intended to match the semantics of the
+  /// MSVC /external:env option.
+  ExternalAfter,
+
   /// Like Angled, but marks system directories.
   System,